Debug pipeline int=>size_t

+++ solvers/CompressibleFlowSolver/ArtificialDiffusion/ArtificialDiffusion.cpp
@@ -120,7 +120,7 @@ void ArtificialDiffusion::v_DoArtificialDiffusion_coeff(
             Array<OneD,       Array<OneD, NekDouble> > &outarray)
 {
     size_t nvariables = inarray.size();
-    int ncoeffs    = m_fields[0]->GetNcoeffs();
+    size_t ncoeffs    = m_fields[0]->GetNcoeffs();
 
     Array<OneD, Array<OneD, NekDouble> > outarrayDiff {nvariables};
